Limit memory allocations instead of buffer acquisitions (#1844)
Move memory limiting to the time of actual heap allocation, rather than
on requesting a buffer. When returning an unused buffers from an already
allocated page, the pool can avoid checking against the limit.

Also included in this change:
* Vocabulary cleanup in code and metrics: buffers handed out to
consumers are now described as "acquired" / "in use" rather than
"reserved" , reserve now means only the prefetch *intent* reservation
(e.g. `mem_reserved`).
* Processing of pending allocations was moved to a background thread.
Releasing a buffer or deallocating memory now only notifies the thread
to trigger processing, rather than invoking it directly, thus avoiding
recursive calls.
* Fallback mechanism to allocate smaller pages in the memory pool when
under memory pressure. Memory pool pages are still created by default
with a buffer count of 16, however if the allocation is blocked by the
memory limiter, will iteratively scaled down by halving the buffer
count, down to a single buffer.
* Merged the functionality of `PoolStats` into `MemoryLimiter`, which
now controls all allocations and deallocations and maintains stats like
`allocated_bytes` by handing out the new low-level type `ManagedBuffer`.

### Does this change impact existing behavior?

N/A

### Does this change need a changelog entry? Does it require a version
change?

N/A

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

+++ mountpoint-s3-fs/examples/mount_from_config.rs
@@ -75,7 +75,7 @@ struct ConfigOptions {
     user_agent_prefix: Option<String>,
     part_size: Option<usize>,
     /// Target memory limit (in bytes) that Mountpoint will try to enforce
-    memory_limit_bytes: Option<u64>,
+    memory_limit_bytes: Option<usize>,
 
     // File system options
     dir_mode: Option<u16>,
